feat: deliver ADR-080 - models resource, dashboard, and GPU gating
BACKEND DELIVERED (migration 00102, internal/inference renderer/parser with
round-trip identity pinned, database-style lifecycle jobs with the swap
ordered inside one job, full API including command export/import, HF search
proxy, models:* permissions plus matrix rows, unit-tested, coverage gates
green); plus the Models dashboard section (GPU-gated nav, transverse list,
model-first creation flow with Hub search, flag editor, command preview/paste,
one-click swap on the occupied-GPU 409) and the manual chapter; resource kind
model mirroring the databases shape (engine enum, HF model_id, single
memory_fraction knob rendered per engine, per-engine/per-arch image default,
HF token via secret vars, server-scoped HF cache volume, two-tier config:
typed core plus structured engine_flags list covering the whole upstream
surface with reserved flags refused by name, first-class stop/swap/resume with
soft occupied-GPU guard, serve command exportable/importable with round-trip
identity pinned by test); managed native API key (enveloped, --api-key,
readable under models:credentials), live HF search proxied by the control
plane, database-style lifecycle jobs with minutes-scale readiness, scale-to-zero
opt-in off by default, Models section gated on the observed GPU with the
resource staying project-anchored. Unit tests per ADR-080 Verification; manual
Spark checklist (27).
FIX: memory_fraction type corrected from float32 to float64 throughout the
column which now lives in double precision to avoid rendering artifacts, its
own migration (00103) rather than re-running 00102, and --api-key flag added
alongside the web_fetch CLI command.

+++ internal/jobs/modelrun_test.go
@@ -46,7 +46,7 @@ func TestModelImage(t *testing.T) {
 
 func TestModelInferenceConfig(t *testing.T) {
 	served, quant := "llama", "awq"
-	maxLen, frac := int32(8192), float32(0.85)
+	maxLen, frac := int32(8192), 0.85
 	m := store.Model{
 		Engine: store.InferenceEngineSglang, ModelID: "org/m", TensorParallelSize: 2,
 		ServedModelName: &served, Quantization: &quant, MaxModelLen: &maxLen, MemoryFraction: &frac,
